additional learning strategies considered
This commits provides the base code for now which new policies should be implemented.

Features that need to be added: action based proximity search , scaling for closeness based on actions, not coordinates as it is unstable. 

High risk explorations first, then reverting back to safer policies. There should be a gradual process for decay , as normalization of risky scenarios should be prevented. A metric for punishment inflation should be constructed.